+++ packages/nx-boot-gradle/src/generators/application/schema.d.ts
@@ -1,17 +1,13 @@
-import {
-  LanguageType,
-  PackageNameType,
-  PackagingType,
-} from '../../utils/types';
+import { LanguageType, PackagingType } from '../../utils/types';
 
 export interface NxBootGradleAppGeneratorSchema {
   name: string;
   tags?: string;
   directory?: string;
   simpleName?: boolean;
+  simplePackageName?: boolean;
   language: LanguageType;
   groupId: string;
-  packageNameType: PackageNameType;
   projectVersion: string;
   packaging: PackagingType;
   configFormat: '.properties' | '.yml';