+++ packages/nx-boot-maven/src/generators/application/schema.d.ts
@@ -1,8 +1,4 @@
-import {
-  LanguageType,
-  PackageNameType,
-  PackagingType,
-} from '../../utils/types';
+import { LanguageType, PackagingType } from '../../utils/types';
 
 export interface NxBootMavenAppGeneratorSchema {
   name: string;
@@ -11,7 +7,6 @@ export interface NxBootMavenAppGeneratorSchema {
   simpleName?: boolean;
   language: LanguageType;
   groupId: string;
-  packageNameType: PackageNameType;
   projectVersion: string;
   packaging: PackagingType;
   configFormat: '.properties' | '.yml';