+++ packages/nx-quarkus-maven/src/generators/application/schema.d.ts
@@ -1,13 +1,13 @@
-import { LanguageType, PackageNameType } from '../../utils/types';
+import { LanguageType } from '../../utils/types';
 
 export interface NxQuarkusMavenAppGeneratorSchema {
   name: string;
   tags?: string;
   directory?: string;
   simpleName?: boolean;
+  simplePackageName?: boolean;
   language: LanguageType;
   groupId: string;
-  packageNameType: PackageNameType;
   projectVersion: string;
   configFormat: '.properties' | '.yml';
   parentProject: string;