+++ packages/nx-quarkus-maven/src/utils/types.ts
@@ -1,4 +1,3 @@
 export type LanguageType = 'java' | 'kotlin';
 export type MvnBuildCommandType = 'compile' | 'package' | 'install';
 export type LinterType = 'checkstyle' | 'pmd' | 'ktlint';
-export type PackageNameType = 'short' | 'long';